feat: added css to view_students (working delete confirmation)

+++ front-end/src/components/pages/View_Students.js
@@ -10,7 +10,7 @@ import '../../css/view_students.css'
 import Header from '../components/Header';
 import Footer from '../components/Footer';
 import Menu from '../components/Menu';
-import Popup from '../components/button_component/Popup';
+import DeletePopup from '../components/Popups/DeletePopup';
 
 const View_Students =()=>{
 
@@ -20,6 +20,8 @@ const [orderValue, setOrderValue] = useState("");
 const [searchValue, setSearchValue] = useState("");
 const [viewValue, setViewValue] = useState("");
 const [input, setInput] = useState("")
+const [showConfirmation, setShowConfirmation] = useState("")
+const [toDelete, setToDelete] = useState("")
 const prev_order_state = useRef();
 const prev_view_state = useRef();
 const orderFilter = [
@@ -50,7 +52,6 @@ prev_order_state.current = [orderValue];
 prev_view_state.current = [viewValue];
 
 const { user, isAuthenticated } = useStore();
-const [isToggled, setToggled] = useState(false)
 
 const navigate = useNavigate();     // navigation hook
 
@@ -71,7 +72,7 @@ const navigate = useNavigate();     // navigation hook
         if(json.result.success){
             setRecord(json.result.output)
         }else{
-            alert(json.result.message)
+            //alert(json.result.message)
         }
     })}
 },[isAuthenticated, state]);
@@ -212,28 +213,35 @@ const handleUserInput = (e) => {
     setInput(value);
 }
 
-const onDelete=async (student)=>{
-    let student_id = student.rec.student_id
-    await fetch('http://localhost:3001/api/0.1/student/'+student_id+'/'+user.user_id,{
-        method: "DELETE",
-        credentials:'include'
-    }).then(response =>{ return response.json()})
-    .then(json=>{
-        console.log(json)
-        if(json.result.success){
-            window.alert(json.result.message)
-            changeState(!state)
-        }
-    })
+const confirmDelete= async(decision) =>{
+    setShowConfirmation(false)
+    if(decision){
+        const student = toDelete.student_id
+        await fetch('http://localhost:3001/api/0.1/student/'+student+'/'+user.user_id,{
+            method: "DELETE",
+            credentials:'include'
+        }).then(response =>{ return response.json()})
+        .then(json=>{
+            if(json.result.success){
+                window.alert(json.result.message)
+                changeState(!state)
+            }
+        })
+    }
+}
+
+const onDelete=(student)=>{
+    setShowConfirmation(true)
+    setToDelete(student);
 }
 
 const DropDown =({value,options,onChange, type})=>{
     return(
         <label>
-            <select value={value} onChange={onChange}>
-                {type === "search"? <option value = "" disabled>search by</option>: type==="view"?  <option value = "" disabled>view by</option>: <option value = "" disabled>order by</option> }
+            <select value={value} onChange={onChange} className='view-student-dropdown'>
+                {type === "search"? <option value = "" disabled hidden>search by</option>: type==="view"?  <option value = "" disabled hidden>view by</option>: <option value = "" disabled hidden>order by</option> }
                 {options.map((option,i)=>(
-                    <option key={i} value = {option.value}>{option.label}</option>
+                    <option key={i} value = {option.value} >{option.label}</option>
                 ))}
             </select>
         </label>
@@ -242,59 +250,58 @@ const DropDown =({value,options,onChange, type})=>{
 
 return(
     <div>
-        <Header />
-        <Menu />
         <div className='view-student-body'>
-    
-            <div className='view-student-header'>
-                <span>Student Records</span>
-                <ul className='view-student-list'>
-                    <li><DropDown options={searchFilter} className='view-student-dropdown' value = {searchValue} onChange={searchChange} type ="search"/></li>
-                    <li><DropDown options={orderFilter} className='view-student-dropdown' value = {orderValue} onChange={orderChange} type = "order"/></li>
-                    <li><DropDown options={viewFilter} className='view-student-dropdown' value = {viewValue} onChange={viewChange} type="view"/></li>
-                </ul>
-            </div>
-            <hr className='view-student-line'/>
+            <ul className='view-student-list'>
+                <li><DropDown options={searchFilter}  value = {searchValue} onChange={searchChange} type ="search"/></li>
+                <li><DropDown options={orderFilter}  value = {orderValue} onChange={orderChange} type = "order"/></li>
+                <li><DropDown options={viewFilter} value = {viewValue} onChange={viewChange} type="view"/></li>
+            </ul>
+            <p className="title">Student Records</p>
+            <hr className='line'></hr>
+                
             <div className='view-student-search'>
-                <input type = "text" className = 'view-student-input' placeholder = "Search by name, student number, or degree program..."
+                <input type = "text" className = 'view-student-input' placeholder = "Search a student record"
                 value = {input} onChange = {handleUserInput} required></input>
-                <a href='#' onClick={handleSubmit} ><BsSearch className='view-student-sicon'/></a>      
+                <a href='#' onClick={handleSubmit} ><BsSearch className='search-icon'/></a>      
             </div>
             <div className='view-student-preview'>
                 {record != undefined ? 
                     <table className='view-student-table'>
                         <thead className='view-student-thead'>
                             <tr>
-                                <th className='view-student-cell-header'>NAME</th>
-                                <th className='view-student-cell-header'>STUDENT NUMBER</th>
-                                <th className='view-student-cell-header'>GWA</th>
-                                <th className='view-student-cell-header'>DEGREE PROGRAM</th>
+                                <th className='name-header'>NAME</th>
+                                <th className='studno-header'>STUDENT NUMBER</th>
+                                <th className='degree-header'>DEGREE PROGRAM</th>
                             </tr>
                         </thead>
                         <tbody className = 'view-student-tbody'>
                             {record.map((rec, i) => {
                                 return (
                                     
-                                    <tr className='view-student-element'>
-                                        <a className = "student-tile" href={'/student/'+rec.student_id}><td className='view-student-cell'>{rec.last_name}, {rec.first_name}{rec.middle_name? ', '+rec.middle_name:""}
+                                    <tr className='view-student-element' href={'/student/'+rec.student_id}>
+                                        <a className = "student-tile" href={'/student/'+rec.student_id}>
+                                        <td className='name-cell'>{rec.last_name}, {rec.first_name}{rec.middle_name? ', '+rec.middle_name:""}
                                         {rec.suffix ? ', ' + rec.suffix : ''}</td></a>
-                                        <td className='view-student-cell'>{rec.student_number}</td>
-                                        <td className='view-student-cell'>{rec.gwa}</td>
-                                        <td className='view-student-cell'>{rec.degree_program}</td>
-                                        <a href='#' onClick={()=>{onDelete({rec})}}><AiFillDelete className='view-student-edit-icon'/></a>
-                                        <button onClick={()=>{setToggled(!isToggled)}}><AiFillEye/></button>
+                                        <td className='studno-cell'>{rec.student_number}</td>
+                                        <td className='degree-cell'>{rec.degree_program}</td>
+                                        <a onClick={()=>{onDelete(rec)}}><AiFillDelete className='view-student-edit-icon'/></a>
+                                        {/* <button onClick={()=>{setToggled(!isToggled)}}><AiFillEye/></button> */}
+                                        
                                     </tr>
                                 );
                             })}
                         </tbody>
+                        {showConfirmation===true? <DeletePopup props={{confirmDelete: confirmDelete.bind()}} />:""}
                     </table>:
                 <div>"No students saved"</div>
                 }
-                {isToggled===true? <Popup/>:""}
+                
 
             </div>
 
         </div>
+        <Header />
+        <Menu />
         <Footer />
     </div>
     );
